+++ changelog.md
@@ -5,7 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## [0.3.0] Unreleased [Github repo]
-# Bump to version 0.3.0
+#### Bump to version 0.3.0
+### Added
+- Pytest with devices
+- `__repr__` that returns device info
+### Fix
+- Base code refactoring
 ## [0.2.4] Unreleased [Github repo]
 ### Added
 - `BLE_DEVICE` for bluetooth low energy devices